requirements.txt
Django==5.0.7
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
factory_boy
python-dotenv
drf-yasg
